+++ app/models/Company.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.{Json, Reads, Writes}
 
-case class Company(inn: Int, name: String) extends Serializable {
+case class Company(inn: Long, name: String) extends Serializable {
 
 }
